Backend/requirements.txt
flask
flask-cors
tensorflow==2.20.0
numpy
pillow
matplotlib